+++ CHANGELOG.md
@@ -7,6 +7,8 @@
     up of default values.
   - Json DOM was parsed multiple times, fixed.
   - Overloaded JsonArtemisSerializer#load - deserialize directly from JsonValue.
+  - When loading entities, each deserialized child is assigned an entity id greater
+    than its parent's.
 
 #### Version: 1.2.1 - 2015-11-23
 - **Fix**: `@PooledWeaver` properly resets primitives and strings back to initial values.